# openapi-main.yaml
openapi: 3.0.0
servers: 
  - url: 'http://localhost:3000' 
info:
  title: OpenAPI Specification
  version: "1"
...
paths:
  /users:
    post:
      tags:
        - Users
      operationId: createUser
      summary: "Create a new user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
        '400':
          description: Invalid input
    get:
      tags:
        - Users
      operationId: listUsers
      summary: "List all users"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of users
  /users/{id}:
    get:
      tags:
        - Users
      operationId: getUserById
      summary: "Get user by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of user to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  name:
                    type: string
                  password:
                    type: string
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "User not found"
    put:
      tags:
        - Users
      operationId: updateUser
      summary: "Update an existing user"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of user to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "User not found"
    delete:
      tags:
        - Users
      operationId: deleteUser
      summary: "Delete a user"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of user to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "User not found"
  /users/login:
    post:
      tags:
        - Users
      operationId: loginUser
      summary: "Login user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User logged in
        '400':
          description: Invalid input
...
  /weapons:
    post:
      tags:
        - Weapons
      operationId: createWeapon
      summary: "Create a new weapon"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                min_damage:
                  type: integer
                max_damage:
                  type: integer
                damage_reduction:
                  type: integer
      responses:
        '201':
          description: Weapon created
        '400':
          description: Invalid input
    get:
      tags:
        - Weapons
      operationId: listWeapons
      summary: "List all weapons"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of weapons
  /weapons/{id}:
    get:
      tags:
        - Weapons
      operationId: getWeaponById
      summary: "Get weapon by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  name:
                    type: string
                  min_damage:
                    type: integer
                  max_damage:
                    type: integer
                  damage_reduction:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Weapon not found"
    put:
      tags:
        - Weapons
      operationId: updateWeapon
      summary: "Update an existing weapon"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                min_damage:
                  type: integer
                max_damage:
                  type: integer
                damage_reduction:
                  type: integer
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Weapon not found"
    delete:
      tags:
        - Weapons
      operationId: deleteWeapon
      summary: "Delete a weapon"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Weapon not found"
  /weapons/stats:
    post:
      tags:
        - Weapons stats
      operationId: createWeaponStats
      summary: "Create a new weapon stats"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                weapon_id:
                  type: integer
                stat_id:
                  type: integer
                value:
                  type: integer
      responses:
        '201':
          description: Weapon stats created
        '400':
          description: Invalid input
    get:
      tags:
        - Weapons stats
      operationId: listWeaponStats
      summary: "List all weapon stats"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of weapon stats
  /weapons/{id}/stats:
    get:
      tags:
        - Weapons stats
      operationId: getWeaponStats
      summary: "Get weapon stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  statId:
                    type: integer
                  value:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Weapon not found"
    post:
      tags:
        - Weapons stats
      summary: "Create a new weapon stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                statId:
                  type: integer
                value:
                  type: integer          
      responses:
        '201':
          description: Weapon stats created
        '400':
          description: Invalid input
  /weapons/{id}/stats/{stat_id}:
    get:
      tags:
        - Weapons stats
      operationId: getWeaponStatsById
      summary: "Get weapon stats by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of weapon stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  statId:
                    type: integer
                  value:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Weapon stats not found"
    put:
      tags:
        - Weapons stats
      operationId: updateWeaponStats
      summary: "Update an existing weapon stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of weapon stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                statId:
                  type: integer
                value:
                  type: integer
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Weapon stats not found"
    delete:
      tags:
        - Weapons stats
      operationId: deleteWeaponStats
      summary: "Delete a weapon stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of weapon to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of weapon stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Weapon stats not found"
...